requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pytz>=2023.3
gspread>=6.0.0
google-auth>=2.23.0